+++ src/errors.rs
@@ -1,5 +1,5 @@
 // Conserve backup system.
-// Copyright 2015-2023 Martin Pool.
+// Copyright 2015-2024 Martin Pool.
 
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -19,7 +19,6 @@ use std::path::PathBuf;
 
 use thiserror::Error;
 
-use crate::blockdir::Address;
 use crate::*;
 
 /// Conserve specific error.
@@ -29,8 +28,21 @@ pub enum Error {
     #[error("Block file {hash:?} corrupt: does not have the expected hash")]
     BlockCorrupt { hash: BlockHash },
 
-    #[error("{address:?} extends beyond decompressed block length {actual_len:?}")]
-    AddressTooLong { address: Address, actual_len: usize },
+    #[error("Referenced block {hash} is missing")]
+    BlockMissing { hash: BlockHash },
+
+    #[error("Block {hash} is too short: actual len {actual_len}, referenced len {referenced_len}")]
+    BlockTooShort {
+        hash: BlockHash,
+        actual_len: usize,
+        referenced_len: usize,
+    },
+
+    #[error("Failed to list blocks: {source}")]
+    ListBlocks {
+        #[source]
+        source: transport::Error,
+    },
 
     #[error("Not a Conserve archive (no CONSERVE header found)")]
     NotAnArchive,
@@ -90,14 +102,14 @@ pub enum Error {
         source: globset::Error,
     },
 
-    #[error("Failed to deserialize json from {:?}", path)]
+    #[error("Failed to deserialize json from {path:?}: {source}")]
     DeserializeJson {
         path: String,
         #[source]
         source: serde_json::Error,
     },
 
-    #[error("Failed to serialize json")]
+    #[error("Failed to serialize json: {source}")]
     SerializeJson {
         #[from]
         source: serde_json::Error,
@@ -109,10 +121,10 @@ pub enum Error {
     #[error("Band not found: {band_id}")]
     BandNotFound { band_id: BandId },
 
-    #[error("Failed to list bands")]
+    #[error("Failed to list bands: {source}")]
     ListBands { source: io::Error },
 
-    #[error("Failed to read source file {:?}", path)]
+    #[error("Failed to read source file {path:?}: {source}")]
     ReadSourceFile { path: PathBuf, source: io::Error },
 
     #[error("Unsupported source file kind: {path:?}")]
@@ -121,13 +133,32 @@ pub enum Error {
     #[error("Unsupported symlink encoding: {path:?}")]
     UnsupportedTargetEncoding { path: PathBuf },
 
-    #[error("Failed to read source tree {:?}", path)]
+    #[error("Failed to read source tree {path:?}: {source}")]
     ListSourceTree { path: PathBuf, source: io::Error },
 
-    #[error("Failed to restore {:?}", path)]
-    Restore { path: PathBuf, source: io::Error },
+    #[error("Failed to restore file {path:?}: {source}")]
+    RestoreFile { path: PathBuf, source: io::Error },
 
-    #[error("Failed to restore modification time on {:?}", path)]
+    #[error("Failed to restore symlink {path:?}: {source}")]
+    RestoreSymlink { path: PathBuf, source: io::Error },
+
+    #[error("Failed to read block content {hash} for {apath}: {source}")]
+    RestoreFileBlock {
+        apath: Apath,
+        hash: BlockHash,
+        source: Box<Error>,
+    },
+
+    #[error("Failed to restore directory {path:?}: {source}")]
+    RestoreDirectory { path: PathBuf, source: io::Error },
+
+    #[error("Failed to restore ownership of {path:?}: {source}")]
+    RestoreOwnership { path: PathBuf, source: io::Error },
+
+    #[error("Failed to restore permissions on {path:?}: {source}")]
+    RestorePermissions { path: PathBuf, source: io::Error },
+
+    #[error("Failed to restore modification time on {path:?}: {source}")]
     RestoreModificationTime { path: PathBuf, source: io::Error },
 
     #[error("Unsupported URL scheme {:?}", scheme)]
@@ -143,7 +174,7 @@ pub enum Error {
         source: io::Error,
     },
 
-    #[error("Failed to set owner of {path:?}")]
+    #[error("Failed to set owner of {path:?}: {source}")]
     SetOwner { source: io::Error, path: PathBuf },
 
     #[error(transparent)]
@@ -170,5 +201,35 @@ impl From<jsonio::Error> for Error {
             }, // conflates serialize/deserialize
             jsonio::Error::Transport { source } => Error::Transport { source },
         }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn block_corrupt_message() {
+        let err = Error::BlockCorrupt {
+            hash: BlockHash::hash_bytes(b"hello"),
+        };
+        assert_eq!(
+            format!("{err}"),
+            "Block file e4cfa39a3d37be31c59609e807970799caa68a19bfaa15135f165085e01d41a65ba1e1b146aeb6bd0092b49eac214c103ccfa3a365954bbbe52f74a2b3620c94 corrupt: does not have the expected hash"
+        );
+    }
+
+    #[test]
+    fn restore_to_nonexistent_directory() {
+        // Hopefully, constructing the error from the `io::ErrorKind` will give consistent
+        // messages across platforms.
+        let err = Error::RestoreFile {
+            path: PathBuf::from("/nonexistent"),
+            source: io::Error::from(io::ErrorKind::NotFound),
+        };
+        assert_eq!(
+            format!("{err}"),
+            "Failed to restore file \"/nonexistent\": entity not found"
+        );
     }
 }