+++ src/index.rs
@@ -25,14 +25,10 @@ use tracing::{debug, debug_span, error};
 
 use crate::compress::snappy::{Compressor, Decompressor};
 use crate::counters::Counter;
-use crate::entry::{EntryValue, KindMeta};
-use crate::kind::Kind;
+use crate::entry::KindMeta;
 use crate::monitor::Monitor;
-use crate::owner::Owner;
 use crate::stats::IndexReadStats;
 use crate::transport::local::LocalTransport;
-use crate::transport::Transport;
-use crate::unix_mode::UnixMode;
 use crate::unix_time::FromUnixAndNanos;
 use crate::*;
 
@@ -154,6 +150,8 @@ impl EntryTrait for IndexEntry {
 
 impl IndexEntry {
     /// Copy the metadata, but not the body content, from another entry.
+    ///
+    /// The result has no blocks.
     pub(crate) fn metadata_from(source: &EntryValue) -> IndexEntry {
         let mtime = source.mtime();
         assert_eq!(
@@ -520,9 +518,8 @@ impl<HI: Iterator<Item = Vec<IndexEntry>>> IndexEntryIter<HI> {
 mod tests {
     use tempfile::TempDir;
 
-    use crate::monitor::collect::CollectMonitor;
+    use crate::monitor::test::TestMonitor;
 
-    use super::transport::local::LocalTransport;
     use super::*;
 
     fn setup() -> (TempDir, IndexWriter) {
@@ -572,7 +569,7 @@ mod tests {
         let (_testdir, mut ib) = setup();
         ib.push_entry(sample_entry("/zzz"));
         ib.push_entry(sample_entry("/aaa"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
     }
 
     #[test]
@@ -580,7 +577,7 @@ mod tests {
     fn index_builder_checks_names() {
         let (_testdir, mut ib) = setup();
         ib.push_entry(sample_entry("../escapecat"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
     }
 
     #[test]
@@ -590,7 +587,7 @@ mod tests {
         let (_testdir, mut ib) = setup();
         ib.push_entry(sample_entry("/again"));
         ib.push_entry(sample_entry("/again"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
     }
 
     #[test]
@@ -599,9 +596,9 @@ mod tests {
     fn no_duplicate_paths_across_hunks() {
         let (_testdir, mut ib) = setup();
         ib.push_entry(sample_entry("/again"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
         ib.push_entry(sample_entry("/again"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
     }
 
     #[test]
@@ -612,7 +609,7 @@ mod tests {
     #[test]
     fn basic() {
         let (testdir, mut ib) = setup();
-        let monitor = CollectMonitor::arc();
+        let monitor = TestMonitor::arc();
         ib.append_entries(&mut vec![sample_entry("/apple"), sample_entry("/banana")]);
         let hunks = ib.finish(monitor.clone()).unwrap();
         assert_eq!(monitor.get_counter(Counter::IndexWrites), 1);
@@ -644,9 +641,9 @@ mod tests {
     fn multiple_hunks() {
         let (testdir, mut ib) = setup();
         ib.append_entries(&mut vec![sample_entry("/1.1"), sample_entry("/1.2")]);
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
         ib.append_entries(&mut vec![sample_entry("/2.1"), sample_entry("/2.2")]);
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
 
         let index_read = IndexRead::open_path(testdir.path());
         let it = index_read.iter_entries();
@@ -677,9 +674,9 @@ mod tests {
     fn iter_hunks_advance_to_after() {
         let (testdir, mut ib) = setup();
         ib.append_entries(&mut vec![sample_entry("/1.1"), sample_entry("/1.2")]);
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
         ib.append_entries(&mut vec![sample_entry("/2.1"), sample_entry("/2.2")]);
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
 
         let index_read = IndexRead::open_path(testdir.path());
         let names: Vec<String> = index_read
@@ -761,13 +758,13 @@ mod tests {
         ib.push_entry(sample_entry("/bar"));
         ib.push_entry(sample_entry("/foo"));
         ib.push_entry(sample_entry("/foobar"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
 
         // Make multiple hunks to test traversal across hunks.
         ib.push_entry(sample_entry("/g01"));
         ib.push_entry(sample_entry("/g02"));
         ib.push_entry(sample_entry("/g03"));
-        ib.finish_hunk(CollectMonitor::arc()).unwrap();
+        ib.finish_hunk(TestMonitor::arc()).unwrap();
 
         // Advance to /foo and read on from there.
         let mut it = IndexRead::open_path(testdir.path()).iter_entries();
@@ -802,9 +799,9 @@ mod tests {
         for i in 0..100_000 {
             ib.push_entry(sample_entry(&format!("/{i:0>10}")));
         }
-        ib.finish_hunk(CollectMonitor::arc())?;
+        ib.finish_hunk(TestMonitor::arc())?;
         // Think about, but don't actually add some files
-        ib.finish_hunk(CollectMonitor::arc())?;
+        ib.finish_hunk(TestMonitor::arc())?;
         let read_index = IndexRead::open_path(testdir.path());
         assert_eq!(read_index.iter_hunks().count(), 1);
         Ok(())