+++ src/test_fixtures.rs
@@ -22,8 +22,7 @@ use std::path::{Path, PathBuf};
 
 use tempfile::TempDir;
 
-use crate::backup::BackupOptions;
-use crate::monitor::collect::CollectMonitor;
+use crate::monitor::test::TestMonitor;
 use crate::*;
 
 /// A temporary archive, deleted when it goes out of scope.
@@ -66,10 +65,10 @@ impl ScratchArchive {
         }
 
         let options = &BackupOptions::default();
-        backup(&self.archive, srcdir.path(), options, CollectMonitor::arc()).unwrap();
+        backup(&self.archive, srcdir.path(), options, TestMonitor::arc()).unwrap();
 
         srcdir.create_file("hello2");
-        backup(&self.archive, srcdir.path(), options, CollectMonitor::arc()).unwrap();
+        backup(&self.archive, srcdir.path(), options, TestMonitor::arc()).unwrap();
     }
 
     pub fn transport(&self) -> &dyn Transport {
@@ -200,4 +199,30 @@ where
         .into_iter()
         .map(|entry| entry.apath().clone().into())
         .collect()
+}
+
+/// On Unix, return the gid of a group that the current user is a member of
+/// and that is not the primary group. This can be used to test chgrp operations.
+///
+/// Panics if there is no such group, or if the list of groups can't be retrieved,
+/// which is always the case on macOS.
+#[cfg(target_os = "linux")]
+pub fn arbitrary_secondary_group() -> u32 {
+    let groups = nix::unistd::getgroups().expect("getgroups");
+    let primary_group = nix::unistd::getgid();
+    let secondary_group = groups
+        .iter()
+        .find(|gid| **gid != primary_group)
+        .expect("secondary group");
+    secondary_group.as_raw()
+}
+
+#[cfg(test)]
+mod test {
+    #[cfg(target_os = "linux")]
+    #[test]
+    fn arbitrary_secondary_group_is_found() {
+        let gid = super::arbitrary_secondary_group();
+        assert!(gid > 0);
+    }
 }