+++ tests/cli_tests/validate.rs
@@ -12,7 +12,7 @@ use serde_json::json;
 use serde_json::{Deserializer, Value};
 use tracing::Level;
 
-use super::run_conserve;
+use crate::run_conserve;
 
 fn read_log_json(path: &Path) -> Vec<serde_json::Value> {
     let json_content = std::fs::read_to_string(path).unwrap();
@@ -77,12 +77,12 @@ fn validate_non_fatal_problems_nonzero_result_and_json_log() {
     let events = read_log_json(log_temp.path());
     dbg!(&events);
     let errors = filter_by_level(&events, Level::ERROR);
+    // TODO: Write errors to json, read that json too.
     assert_eq!(errors.len(), 1);
     assert_eq!(
         errors[0]["fields"],
         json!({
-            "block_hash": "fec91c70284c72d0d4e3684788a90de9338a5b2f47f01fedbe203cafd68708718ae5672d10eca804a8121904047d40d1d6cf11e7a76419357a9469af41f22d01",
-            "message": "Referenced block missing",
+            "message": "Referenced block fec91c70284c72d0d4e3684788a90de9338a5b2f47f01fedbe203cafd68708718ae5672d10eca804a8121904047d40d1d6cf11e7a76419357a9469af41f22d01 is missing",
         })
     );
 }