+++ tests/old_archives.rs
@@ -1,5 +1,5 @@
 // Conserve backup system.
-// Copyright 2020, 2022 Martin Pool.
+// Copyright 2020-2023 Martin Pool.
 
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -21,7 +21,8 @@ use std::sync::Arc;
 
 use assert_fs::prelude::*;
 use assert_fs::TempDir;
-use conserve::monitor::collect::CollectMonitor;
+use conserve::counters::Counter;
+use conserve::monitor::test::TestMonitor;
 use predicates::prelude::*;
 use pretty_assertions::assert_eq;
 
@@ -29,15 +30,14 @@ use conserve::*;
 use time::OffsetDateTime;
 use tracing_test::traced_test;
 
+mod util;
+use util::{copy_testdata_archive, testdata_archive_path};
+
 const MINIMAL_ARCHIVE_VERSIONS: &[&str] = &["0.6.0", "0.6.10", "0.6.2", "0.6.3", "0.6.9", "0.6.17"];
 
 fn open_old_archive(ver: &str, name: &str) -> Archive {
-    Archive::open_path(Path::new(&archive_testdata_path(name, ver)))
+    Archive::open_path(Path::new(&testdata_archive_path(name, ver)))
         .expect("Failed to open archive")
-}
-
-fn archive_testdata_path(name: &str, ver: &str) -> String {
-    format!("testdata/archive/{name}/v{ver}/")
 }
 
 #[test]
@@ -83,7 +83,7 @@ fn validate_archive() {
         let archive = open_old_archive(ver, "minimal");
 
         archive
-            .validate(&ValidateOptions::default(), Arc::new(CollectMonitor::new()))
+            .validate(&ValidateOptions::default(), Arc::new(TestMonitor::new()))
             .expect("validate archive");
         assert!(!logs_contain("ERROR") && !logs_contain("WARN"));
     }
@@ -101,16 +101,18 @@ fn long_listing_old_archive() {
         let mut stdout = Vec::<u8>::new();
 
         // show archive contents
+        let monitor = TestMonitor::arc();
         show::show_entry_names(
             archive
                 .open_stored_tree(BandSelectionPolicy::Latest)
                 .unwrap()
-                .iter_entries(Apath::root(), Exclude::nothing())
+                .iter_entries(Apath::root(), Exclude::nothing(), monitor.clone())
                 .unwrap(),
             &mut stdout,
             true,
         )
         .unwrap();
+        monitor.assert_no_errors();
 
         if first_with_perms.matches(&semver::Version::parse(ver).unwrap()) {
             assert_eq!(
@@ -141,18 +143,19 @@ fn restore_old_archive() {
         println!("restore {} to {:?}", ver, dest.path());
 
         let archive = open_old_archive(ver, "minimal");
-        let restore_stats = restore(
+        let monitor = TestMonitor::arc();
+        restore(
             &archive,
             dest.path(),
             &RestoreOptions::default(),
-            CollectMonitor::arc(),
+            monitor.clone(),
         )
         .expect("restore");
 
-        assert_eq!(restore_stats.files, 2);
-        assert_eq!(restore_stats.symlinks, 0);
-        assert_eq!(restore_stats.directories, 2);
-        assert_eq!(restore_stats.errors, 0);
+        monitor.assert_counter(Counter::Symlinks, 0);
+        monitor.assert_counter(Counter::Files, 2);
+        monitor.assert_counter(Counter::Dirs, 2);
+        monitor.assert_no_errors();
 
         dest.child("hello").assert("hello world\n");
         dest.child("subdir").assert(predicate::path::is_dir());
@@ -198,18 +201,13 @@ fn restore_modify_backup() {
             &archive,
             working_tree.path(),
             &RestoreOptions::default(),
-            CollectMonitor::arc(),
+            TestMonitor::arc(),
         )
         .expect("restore");
 
         // Write back into a new copy of the archive, without modifying the
         // testdata in the source tree.
-        let new_archive_temp = TempDir::new().unwrap();
-        let stored_archive_path = archive_testdata_path("minimal", ver);
-        let new_archive_path = new_archive_temp.path().join("archive");
-        cp_r::CopyOptions::default()
-            .copy_tree(stored_archive_path, &new_archive_path)
-            .expect("copy archive tree");
+        let archive_temp = copy_testdata_archive("minimal", ver);
 
         working_tree
             .child("empty")
@@ -221,7 +219,7 @@ fn restore_modify_backup() {
         )
         .expect("overwrite file");
 
-        let new_archive = Archive::open_path(&new_archive_path).expect("Open new archive");
+        let new_archive = Archive::open_path(archive_temp.path()).expect("Open new archive");
         let emitted = RefCell::new(Vec::new());
         let backup_stats = backup(
             &new_archive,
@@ -235,7 +233,7 @@ fn restore_modify_backup() {
                 })),
                 ..Default::default()
             },
-            CollectMonitor::arc(),
+            TestMonitor::arc(),
         )
         .expect("Backup modified tree");
 
@@ -281,6 +279,6 @@ fn restore_modify_backup() {
         assert_eq!(backup_stats.errors, 0);
 
         working_tree.close().expect("Cleanup working tree");
-        new_archive_temp.close().expect("Cleanup copied archive");
+        archive_temp.close().expect("Cleanup copied archive");
     }
 }