+++ NEWS.md
@@ -1,5 +1,15 @@
 # Conserve release history
 
+## 24.8.0
+
+- Fixed: `restore --only` specifying a subdirectory no longer fails due to parent directories missing from the destination.
+
+- Fixed: More detail about the causes of errors in the log.
+
+- `restore` no longer prints stats, due to internal changes; this will be restored later.
+
+- Minimum Rust version increased to 1.74 due to updated dependencies.
+
 ## 23.11.0
 
 - Fixed: Restore now sets Unix user/group ownership on symlinks and directories. Previously, only file ownership was restored. (Setting file ownership typically requires restoring as root.)